requirements.txt
pyyaml
python-telegram-bot
tencentcloud-sdk-python
pymysql
